remove huggingface datasets from default config
