+++ src/refiners/training_utils/huggingface_datasets.py
@@ -1,6 +1,7 @@
 from typing import Any, Generic, Protocol, TypeVar, cast
 
 from datasets import VerificationMode, load_dataset as _load_dataset  # type: ignore
+from pydantic import BaseModel  # type: ignore
 
 __all__ = ["load_hf_dataset", "HuggingfaceDataset"]
 
@@ -22,3 +23,14 @@ def load_hf_dataset(
     verification_mode = VerificationMode.BASIC_CHECKS if use_verification else VerificationMode.NO_CHECKS
     dataset = _load_dataset(path=path, revision=revision, split=split, verification_mode=verification_mode)
     return cast(HuggingfaceDataset[Any], dataset)
+
+
+class HuggingfaceDatasetConfig(BaseModel):
+    hf_repo: str = "finegrain/unsplash-dummy"
+    revision: str = "main"
+    split: str = "train"
+    horizontal_flip: bool = False
+    random_crop: bool = True
+    use_verification: bool = False
+    resize_image_min_size: int = 512
+    resize_image_max_size: int = 576
